// TsogosunProfileAdminWeb/src/app/model/igt-gaming-system/address.ts
export interface Address {
    address1?: string | undefined;
    city?: string | undefined;
    stateProvince?: string | undefined;
    postalCode?: string | undefined;
    country?: string | undefined;
    location?: string | undefined;
    deliverable?: string | undefined;
    preferred?: string | undefined;
    suburb?: string | undefined;
    creditAddress?: string | undefined;
}